+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
   - `speed`, which defines how many world units the lift will travel per frame
 - added properties to the `O_BIG_BOWL` object:
   - `pour_time`, which defines how long to pour liquid from the bowl
+  - `flip_slot`, which defines the flip map slot index to alter once liquid has finished pouring
 - changed `Game mode selection` to allow hiding the dialog even after having completed the game (Gameplay → General → Game mode selection)
 - improved handling of dead enemies used as switch triggers in custom levels by not altering their activation status and by adding detection for having been exploded (#5682)
 - fixed Lara having an empty left holster if the option to remember guns between levels is used and she finished with the Desert Eagle (#5697)